Add co_post network type
This indicates any kind of co-temporal messaging behaviour, regardless
of whether it uses specific platform features. This kind of network
indicates the maximum possible weight between any two accounts for
a specific time window, so it may not be directly useful except
as a normalisation factor.

Also automatically use all of the available CPU for computing network
types.