Exclude comparison of a message with itself from calculations
Changes the weight quantification to exclude comparisons of a
message with itself. Previous to this change every activity by
a user that met the criteria for that network type would be
included in the self edge weight - eg. every retweet would be
a co-retweet with itself.

The previous behaviour is undesirable because it significantly
inflates the weight of self edges according to the account
activity - this change gives a more accurate representation
of the self edge weight, allowing easier detection of things
like reply spam from a single account.